Move colorbar to colorbar.py
