+++ pygmt/src/__init__.py
@@ -4,6 +4,7 @@ Source code for PyGMT modules.
 # pylint: disable=import-outside-toplevel
 # GMT Primary modules
 from pygmt.src.coast import coast
+from pygmt.src.colorbar import colorbar
 from pygmt.src.logo import logo
 
 # GMT Supplementary modules
